+++ Classes/Tool/Tca/Builder/FieldPreset/AbstractFieldPreset.php
@@ -436,11 +436,11 @@ abstract class AbstractFieldPreset implements FieldPresetInterface
      *
      * @param   array       $optionDefinition
      * @param   array|null  $type
-     * @param   string      $default
+     * @param   mixed       $default
      *
      * @return array
      */
-    protected function addDefaultOptions(array $optionDefinition, ?array $type = null, string $default = ''): array
+    protected function addDefaultOptions(array $optionDefinition, ?array $type = null, $default = ''): array
     {
         $optionDefinition['default'] = [
             'type' => $type ?? ['string'],
@@ -558,9 +558,11 @@ abstract class AbstractFieldPreset implements FieldPresetInterface
             $this->configureSqlColumn(static function (Column $column) use ($options) {
                 if ((int)$options['maxLength'] <= 4096) {
                     $column->setType(new StringType())
+                           ->setDefault('')
                            ->setLength((int)$options['maxLength']);
                 } else {
                     $column->setType(new TextType())
+                           ->setDefault(null)
                            ->setLength(null);
                 }
             });