+++ ext_emconf.php
@@ -1,22 +1,4 @@
 <?php
-/*
- * Copyright 2021 LABOR.digital
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Last modified: 2021.06.27 at 16:32
- */
-
 declare(strict_types=1);
 /*
  * Copyright 2021 LABOR.digital
@@ -54,7 +36,7 @@ $EM_CONF[$_EXTKEY] = [
     'modify_tables' => '',
     'clearCacheOnLoad' => 1,
     'lockType' => '',
-    'version' => "10.1.2",
+    'version' => "10.4.0",
     'constraints' => [
         'depends' => [
             'typo3' => '10.0.0-10.99.99',